+++ src/main.h
@@ -131,8 +131,9 @@ public:
     }
 
     IMPLEMENT_SERIALIZE( READWRITE(FLATDATA(*this)); )
-    void SetNull() { nFile = -1; nBlockPos = 0; nTxPos = 0; }
-    bool IsNull() const { return (nFile == -1); }
+
+    void SetNull() { nFile = static_cast<unsigned int>(-1); nBlockPos = 0; nTxPos = 0; }
+    bool IsNull() const { return (nFile == static_cast<unsigned int>(-1)); }
 
     friend bool operator==(const CDiskTxPos& a, const CDiskTxPos& b)
     {
@@ -171,8 +172,9 @@ public:
 
     CInPoint() { SetNull(); }
     CInPoint(CTransaction* ptxIn, unsigned int nIn) { ptx = ptxIn; n = nIn; }
-    void SetNull() { ptx = NULL; n = -1; }
-    bool IsNull() const { return (ptx == NULL && n == -1); }
+
+    void SetNull() { ptx = NULL; n = static_cast<unsigned int>(-1); }
+    bool IsNull() const { return (ptx == NULL && n == static_cast<unsigned int>(-1)); }
 };
 
 
@@ -186,9 +188,11 @@ public:
 
     COutPoint() { SetNull(); }
     COutPoint(uint256 hashIn, unsigned int nIn) { hash = hashIn; n = nIn; }
+
     IMPLEMENT_SERIALIZE( READWRITE(FLATDATA(*this)); )
-    void SetNull() { hash = 0; n = -1; }
-    bool IsNull() const { return (hash == 0 && n == -1); }
+
+    void SetNull() { hash = 0; n = static_cast<unsigned int>(-1); }
+    bool IsNull() const { return (hash == 0 && n == static_cast<unsigned int>(-1)); }
 
     friend bool operator<(const COutPoint& a, const COutPoint& b)
     {
