+++ src/script.cpp
@@ -1085,7 +1085,7 @@ bool EvalScript(vector<vector<unsigned char> >& stack, const CScript& script, co
                 {
                     // ([sig ...] num_of_signatures [pubkey ...] num_of_pubkeys -- bool)
 
-                    int i = 1;
+                    size_t i = 1;
                     if (stack.size() < i)
                         return false;
